fix factory

+++ MsdProjectAuthService/src/main/java/com/bah/msd/domain/CustomerFactory.java
@@ -7,7 +7,7 @@ public class CustomerFactory {
 	public static Customer getCustomer(String jsonString){
         JSONObject jsonObject = new org.json.JSONObject(jsonString); 
           
-        long id = (long) jsonObject.get("id");
+        int id = (int) jsonObject.get("id");
         String name = (String) jsonObject.get("name"); 
         String email = (String) jsonObject.get("email"); 
         String password = (String) jsonObject.get("password");
